requirements.txt
# API
fastapi==0.68.1
uvicorn[standard]==0.15.0
Werkzeug==2.0.2

# DB
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.6.5

# Tests
pytest==6.2.4
pytest-cov==2.12.1
factory_boy==3.2.0
requests==2.25.1